refactor(review): ReviewService → CommandService 구조로 리팩토링 및 인터페이스 분리
