+++ src/main/java/com/example/umc9th/domain/review/service/command/ReviewCommandService.java
@@ -1,0 +1,5 @@
+package com.example.umc9th.domain.review.service.command;
+
+public interface ReviewCommandService {
+    void createReview();
+}
